Renamed event to OnDatapackSyncEvent. Moved it to fire before tags sync on /reload. Updated test mod
